requirements.txt
gradio>=4.0.0
openai>=1.0.0
psutil>=5.9.0
pydantic>=2.0.0
pywin32>=306; sys_platform == "win32"
python-dotenv>=1.0.0
typing-extensions>=4.5.0 